Fixed not equal folding
The not equal folding logic was wrong. It would end the loop
as soon as it found equal components, which was checking if
ALL components are not equal, but that's not how not equal
works, it should return true if ANY component is not equal.

I also refactored it because it was hard to follow, with the
boolNodeFlag variable being used upside down of the result.

Change-Id: I704f26f2fd31a1cc637f5d7601409de8161c5b80
Reviewed-on: https://swiftshader-review.googlesource.com/5211
Tested-by: Alexis Hétu <sugoi@google.com>
Reviewed-by: Nicolas Capens <capn@google.com>