Fix Mac CMake build.
On Mac the linker wants to determine the full vtable if any virtual methods
has been defined (even when no objects are instantiated). We don't have the
definitions for all the virtual methods for objects defined in the other
library. Work around it by defining the key methods as inline.

Bug swiftshader:41

Change-Id: I9f8b4e98962649dd92f9f6e84e6474f921c4b9e9